separated spin-orbital, closed-shell spintrace and open-shell spintrace examples
